Rename the component folder to match what callers use
