+++ common/views/components/TextWithDot/Dot.tsx
