+++ common/views/components/TextWithDot/index.tsx
