+++ content/webapp/components/StatusIndicator/StatusIndicator.tsx
@@ -1,6 +1,6 @@
 import { font } from '@weco/common/utils/classnames';
 import { FunctionComponent } from 'react';
-import TextWithDot from '@weco/common/views/components/Dot/TextWithDot';
+import TextWithDot from '@weco/common/views/components/TextWithDot';
 import {
   addDays,
   isFuture,
